integer fixed

+++ player.py
@@ -7,7 +7,7 @@ class Player:
     def betRequest(self, game_state):
         our_player = development.get_our_info()
         if development.is_high_cards(our_player):
-            return max(game_state["game_state"]["current_buy_in"], our_player["stack"]*0.1)
+            return max(int(game_state["game_state"]["current_buy_in"]), int(our_player["stack"]*0.1))
         return 0
 
     def showdown(self, game_state):
